namespace adjustments
