+++ Drakengard1and2Extractor.csproj
@@ -204,14 +204,14 @@
       <SubType>Component</SubType>
     </Compile>
     <Compile Include="Drakengard1and2Extractor\Support\CommonMethods.cs" />
-    <Compile Include="Drakengard1and2Extractor\Support\Lz0Decompression.cs" />
-    <Compile Include="Drakengard1and2Extractor\Support\ImageHelpers\BmpPngHelpers.cs" />
-    <Compile Include="Drakengard1and2Extractor\Support\ImageHelpers\DDSimgHelpers.cs" />
-    <Compile Include="Drakengard1and2Extractor\Support\ImageHelpers\UnSwizzleHelpers.cs" />
+    <Compile Include="Drakengard1and2Extractor\Support\Lz0Helpers\Lz0Decompression.cs" />
+    <Compile Include="Drakengard1and2Extractor\Support\ImageHelpers\BmpPngExporter.cs" />
+    <Compile Include="Drakengard1and2Extractor\Support\ImageHelpers\DDSimgExporter.cs" />
+    <Compile Include="Drakengard1and2Extractor\Support\ImageHelpers\PS2UnSwizzlers.cs" />
     <Compile Include="Drakengard1and2Extractor\Support\ImageHelpers\ImgOptions.cs" />
     <Compile Include="Drakengard1and2Extractor\Support\LoggingHelpers.cs" />
-    <Compile Include="Drakengard1and2Extractor\Support\Minilz0Helpers.cs" />
-    <Compile Include="Drakengard1and2Extractor\Support\Extensions\StreamHelpers.cs" />
+    <Compile Include="Drakengard1and2Extractor\Support\Lz0Helpers\Minilz0Function.cs" />
+    <Compile Include="Drakengard1and2Extractor\Support\Extensions\StreamExtension.cs" />
     <Compile Include="Drakengard1and2Extractor\BinExtraction\Drk1BIN.cs" />
     <Compile Include="Drakengard1and2Extractor\BinExtraction\Drk2BIN.cs" />
     <Compile Include="Drakengard1and2Extractor\ImageConversion\FileSPK0.cs">
